chore: stop emmitting sourcemaps in integ-tests (#29978)
### Reason for this change

While working on a related issue, I noticed that some integ-test snapshots are very brittle because assets include inline sourcemaps and ts files.

### Description of changes

With this change, we are disabling all sourcemaps for integ-tests and explicitly exclude ts source files from assets.
This is okay because the handlers are just written to be used inside integ tests.

Another change was made to disable the inline sourcemap for the custom resource handler entrypoint. The actual handlers do not have sourcemaps, so this was silly and probably just an oversight.

### Description of how you validated changes

Run integ tests to update snapshots.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*