feat(cloudfront): add `autoPublish` property to the `Function` (#30030)
### Issue # (if applicable)

Closes #30028 

### Reason for this change

As described in the issue.

### Description of changes


In this pull request, the "autoPublish" Property has been added to the Function class.

If you set `autoPublish` to false, the function will not be automatically published to the LIVE stage when it’s created.

```ts
new cloudfront.Function(this, 'Function', {
  code: cloudfront.FunctionCode.fromInline('function handler(event) { return event.request }'),
  runtime: cloudfront.FunctionRuntime.JS_2_0,
  autoPublish: false // Add
});
```

To ensure that the behavior remains unchanged when autoPublish is not set, the field is optional, with the default value set to true.


### Description of how you validated changes


I've added unit and integ tests.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*