chore(lambda-nodejs): clean up sdk v2 references and warn about runtime updates related to sdk bundling (#30099)
Closes #29836.

While sdk v2 was not being used directly in the construct, we had some remnants of its use. Added a warning for users of Node 16 who do not bundle the sdk on what it will take to update to newer node versions. Also updated integ tests to make actual sdk v3 calls as well as clean up false references to v2.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*