chore(region-info): add missing dlc repository accounts. (#30147)
### Issue # (if applicable)

Add missing DLC repository account information for multiple regions.

### Reason for this change
As part of exteding support to new regions we require the `DLC_REPOSITORY_ACCOUNTS` for those specific regions in our `fact-tables`.


### Description of changes

Added missing information for few regions we already provide support to. Also, updated the reference doc links since they are out of date and haven't been updated in a long time and we need to reference the new docs to fetch the required information.

### Description of how you validated changes



### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*