feat(cloudwatch): add support for labels on pie graphs (aws#28929).
Add a new "labels" property for graph widgets to support toggling visibility status.

=== Issue #28929 ===

Closes #28929.

=== Reason for this change ===

The labels property that allows for toggling the visibility of labels on graphs is currently missing.

=== Description of changes ===

The commits adds a new `"labels"` property with the type `LabelProps` on the graph widget.
The `LabelProps` type contains a single boolean key`"visible"` which toggles the presence of the label on a pie graph. When set to `true`, CloudWatch dashboard will render the percentage ratios of each segment of the pie graph.
The README file has been updated to reflect the new property and how it can be used.

=== Description of how you validated changes ===

The commit includes unit, integration, and snapshot tests as would be generated by the CDK library.
The change has also been manually verified from the CloudWatch console by enabling and disabling the property.

=== Checklist ===

- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*