fix(fsx): no validation error when `hour` property set to 24 in the `LustreMaintenanceTime` class. (#30342)
### Issue # (if applicable)

Closes #30341 

### Reason for this change
The `hour` property in the `LustreMaintenanceTime` class should be between 0 and 23.

But no validation error occurs when it is set to 24.



### Description of changes
In the validate method, I changed it so that an error is thrown when the hour is greater than 23, instead of when the hour is greater than 24. 

This allows a validation error to occur when the hour is set to 24.



### Description of how you validated changes
Changed unit tests and add integ tests.


### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*