chore(lambda): match enum value to property (#30472)
Looks like we are comparing `LogFormat` to the `LoggingFormat` enum. This is not technically a problem for us because `LogFormat` is a string. We deprecated it and created `LoggingFormat` to enforce the enum.

This conditional works because all the enums collapse to strings, but it is best if we sync up the properties with their correct enums.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*