chore: upgrade jsii to `~5.4.21` (#30518)
### Reason for this change

There was an upstream bug in jsii https://github.com/aws/jsii-compiler/issues/1070 that caused the build to fail.
Updating the package manually to confirm the issue is resolved.

### Description of changes

Upgraded to the latest version of jsii that includes a fix for the build issue.

### Description of how you validated changes

Dependency upgrade. Existing tests cover this.

### Checklist
- [x] My code adheres to the [CONTRIBUTING GUIDE](https://github.com/aws/aws-cdk/blob/main/CONTRIBUTING.md) and [DESIGN GUIDELINES](https://github.com/aws/aws-cdk/blob/main/docs/DESIGN_GUIDELINES.md)

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*